fix(injector): allow using static injector without transformers
