+++ lib/di_static.dart
@@ -1,0 +1,7 @@
+library di_static;
+
+export 'key.dart' show Key, key;
+export 'src/injector.dart' show Injector, ModuleInjector;
+export 'src/module_static.dart' show Module, Binding, DEFAULT_VALUE;
+export 'src/reflector.dart' show TypeReflector;
+export 'src/errors.dart' hide BaseError, PRIMITIVE_TYPES;
